ml-worker/embedder/requirements.txt
# requirements.txt — embedder

pika==1.3.2
pinecone-client==3.1.0
sentence-transformers==2.2.2
# ไม่ต้องระบุ torch / transformers / huggingface-hub
# sentence-transformers จะจัดการ deps เองอัตโนมัติ